@@ -25,6 +25,9 @@
     "url": "https://github.com/springload/draftail/issues"
   },
   "homepage": "https://github.com/springload/draftail#readme",
+  "eslintConfig": {
+    "extends": "springload"
+  },
   "jest": {
     "testPathIgnorePatterns": [
       "/node_modules/",
@@ -48,28 +51,30 @@
     "coveralls": "^2.11.15",
     "dotenv": "^4.0.0",
     "draft-js": "^0.10.0",
-    "enzyme": "^2.7.0",
-    "enzyme-to-json": "^1.4.5",
-    "eslint": "^3.13.1",
-    "eslint-config-airbnb": "^14.0.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.9.0",
+    "enzyme": "^2.8.2",
+    "enzyme-to-json": "^1.5.1",
+    "eslint": "3.19.0",
+    "eslint-config-springload": "1.0.2",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "3.0.2",
+    "eslint-plugin-react": "6.10.3",
     "express": "^4.14.0",
     "jest": "^19.0.2",
     "node-sass": "^4.3.0",
     "progress-bar-webpack-plugin": "^1.9.0",
-    "react": "^15.4.1",
-    "react-addons-test-utils": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-test-renderer": "^15.5.4",
     "webpack": "^1.13.3",
     "webpack-dev-middleware": "^1.8.4",
     "webpack-hot-middleware": "^2.13.2"
   },
   "peerDependencies": {
     "draft-js": "^0.10.0",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0"
+    "prop-types": "^15.5.0",
+    "react": "^15.5.0",
+    "react-dom": "^15.5.0"
   },
   "scripts": {
     "start": "npm run server -s",
