@@ -1,6 +1,8 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import { AtomicBlockUtils } from 'draft-js';
 
+/* global EMBEDLY_API_KEY */
 const EMBEDLY_ENDPOINT = `https://api.embedly.com/1/oembed?key=${EMBEDLY_API_KEY}`;
 
 const getJSON = (endpoint, data, successCallback) => {
@@ -47,9 +49,9 @@ class EmbedSource extends React.Component {
 }
 
 EmbedSource.propTypes = {
-    editorState: React.PropTypes.object.isRequired,
-    options: React.PropTypes.object.isRequired,
-    onUpdate: React.PropTypes.func.isRequired,
+    editorState: PropTypes.object.isRequired,
+    options: PropTypes.object.isRequired,
+    onUpdate: PropTypes.func.isRequired,
 };
 
 export default EmbedSource;
