@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 import { NBSP } from '../api/constants';
@@ -38,9 +39,9 @@ const EmbedBlock = ({ entity, isActive, onClick }) => {
 };
 
 EmbedBlock.propTypes = {
-    entity: React.PropTypes.object.isRequired,
-    isActive: React.PropTypes.bool.isRequired,
-    onClick: React.PropTypes.func.isRequired,
+    entity: PropTypes.object.isRequired,
+    isActive: PropTypes.bool.isRequired,
+    onClick: PropTypes.func.isRequired,
 };
 
 export default EmbedBlock;
