@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 
 import Icon from '../components/Icon';
@@ -130,12 +131,12 @@ class ImageBlock extends Component {
 }
 
 ImageBlock.propTypes = {
-    entityConfig: React.PropTypes.object.isRequired,
-    entity: React.PropTypes.object.isRequired,
-    isActive: React.PropTypes.bool.isRequired,
-    onClick: React.PropTypes.func.isRequired,
-    onSave: React.PropTypes.func.isRequired,
-    onCancel: React.PropTypes.func.isRequired,
+    entityConfig: PropTypes.object.isRequired,
+    entity: PropTypes.object.isRequired,
+    isActive: PropTypes.bool.isRequired,
+    onClick: PropTypes.func.isRequired,
+    onSave: PropTypes.func.isRequired,
+    onCancel: PropTypes.func.isRequired,
 };
 
 export default ImageBlock;
