@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 
 import { ENTITY_TYPE } from '../api/constants';
@@ -75,9 +76,9 @@ class MediaBlock extends Component {
 }
 
 MediaBlock.propTypes = {
-    block: React.PropTypes.object.isRequired,
-    contentState: React.PropTypes.object.isRequired,
-    blockProps: React.PropTypes.object.isRequired,
+    block: PropTypes.object.isRequired,
+    contentState: PropTypes.object.isRequired,
+    blockProps: PropTypes.object.isRequired,
 };
 
 export default MediaBlock;
