@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 import Icon from '../components/Icon';
@@ -23,11 +24,11 @@ const Button = ({ active, label, title, icon, onClick }) => (
 );
 
 Button.propTypes = {
-    active: React.PropTypes.bool,
-    label: React.PropTypes.string,
-    title: React.PropTypes.string,
-    icon: React.PropTypes.string,
-    onClick: React.PropTypes.func,
+    active: PropTypes.bool,
+    label: PropTypes.string,
+    title: PropTypes.string,
+    icon: PropTypes.string,
+    onClick: PropTypes.func,
 };
 
 Button.defaultProps = {
