@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 
 import {
@@ -49,38 +50,38 @@ const defaultProps = {
 };
 
 const propTypes = {
-    rawContentState: React.PropTypes.object,
-    onSave: React.PropTypes.func,
-    enableHorizontalRule: React.PropTypes.bool,
-    enableLineBreak: React.PropTypes.bool,
-    stripPastedStyles: React.PropTypes.bool,
-    entityTypes: React.PropTypes.arrayOf(React.PropTypes.shape({
-        label: React.PropTypes.string.isRequired,
-        type: React.PropTypes.string.isRequired,
-        icon: React.PropTypes.string,
-        control: React.PropTypes.func.isRequired,
-        strategy: React.PropTypes.func,
-        component: React.PropTypes.func,
+    rawContentState: PropTypes.object,
+    onSave: PropTypes.func,
+    enableHorizontalRule: PropTypes.bool,
+    enableLineBreak: PropTypes.bool,
+    stripPastedStyles: PropTypes.bool,
+    entityTypes: PropTypes.arrayOf(PropTypes.shape({
+        label: PropTypes.string.isRequired,
+        type: PropTypes.string.isRequired,
+        icon: PropTypes.string,
+        control: PropTypes.func.isRequired,
+        strategy: PropTypes.func,
+        component: PropTypes.func,
         // A bit dodgy. Specific to the IMAGE type or whatever uses the ImageBlock.
-        imageFormats: React.PropTypes.arrayOf(React.PropTypes.shape({
-            label: React.PropTypes.string.isRequired,
-            value: React.PropTypes.string.isRequired,
+        imageFormats: PropTypes.arrayOf(PropTypes.shape({
+            label: PropTypes.string.isRequired,
+            value: PropTypes.string.isRequired,
         })),
     })),
-    blockTypes: React.PropTypes.arrayOf(React.PropTypes.shape({
-        label: React.PropTypes.string.isRequired,
-        type: React.PropTypes.string.isRequired,
-        icon: React.PropTypes.string,
-        element: React.PropTypes.string,
-        className: React.PropTypes.string,
+    blockTypes: PropTypes.arrayOf(PropTypes.shape({
+        label: PropTypes.string.isRequired,
+        type: PropTypes.string.isRequired,
+        icon: PropTypes.string,
+        element: PropTypes.string,
+        className: PropTypes.string,
     })),
-    inlineStyles: React.PropTypes.arrayOf(React.PropTypes.shape({
-        label: React.PropTypes.string.isRequired,
-        type: React.PropTypes.string.isRequired,
-        icon: React.PropTypes.string,
+    inlineStyles: PropTypes.arrayOf(PropTypes.shape({
+        label: PropTypes.string.isRequired,
+        type: PropTypes.string.isRequired,
+        icon: PropTypes.string,
     })),
-    maxListNesting: React.PropTypes.number,
-    stateSaveInterval: React.PropTypes.number,
+    maxListNesting: PropTypes.number,
+    stateSaveInterval: PropTypes.number,
 };
 
 /**
