@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
@@ -31,7 +32,7 @@ class Portal extends React.Component {
 }
 
 Portal.propTypes = {
-    children: React.PropTypes.node,
+    children: PropTypes.node,
 };
 
 Portal.defaultProps = {
