@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 import DraftUtils from '../api/DraftUtils';
@@ -71,17 +72,17 @@ const Toolbar = ({
 );
 
 Toolbar.propTypes = {
-    editorState: React.PropTypes.object.isRequired,
-    enableHorizontalRule: React.PropTypes.bool.isRequired,
-    enableLineBreak: React.PropTypes.bool.isRequired,
-    entityTypes: React.PropTypes.array.isRequired,
-    blockTypes: React.PropTypes.array.isRequired,
-    inlineStyles: React.PropTypes.array.isRequired,
-    toggleBlockType: React.PropTypes.func.isRequired,
-    toggleInlineStyle: React.PropTypes.func.isRequired,
-    addHR: React.PropTypes.func.isRequired,
-    addBR: React.PropTypes.func.isRequired,
-    onRequestDialog: React.PropTypes.func.isRequired,
+    editorState: PropTypes.object.isRequired,
+    enableHorizontalRule: PropTypes.bool.isRequired,
+    enableLineBreak: PropTypes.bool.isRequired,
+    entityTypes: PropTypes.array.isRequired,
+    blockTypes: PropTypes.array.isRequired,
+    inlineStyles: PropTypes.array.isRequired,
+    toggleBlockType: PropTypes.func.isRequired,
+    toggleInlineStyle: PropTypes.func.isRequired,
+    addHR: PropTypes.func.isRequired,
+    addBR: PropTypes.func.isRequired,
+    onRequestDialog: PropTypes.func.isRequired,
 };
 
 export default Toolbar;
