@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 
 import { truncateURL } from '../utils/format';
@@ -40,13 +41,13 @@ const Tooltip = ({ position, entityData, onEdit, onRemove }) => (
 );
 
 Tooltip.propTypes = {
-    position: React.PropTypes.shape({
-        top: React.PropTypes.number.isRequired,
-        left: React.PropTypes.number.isRequired,
+    position: PropTypes.shape({
+        top: PropTypes.number.isRequired,
+        left: PropTypes.number.isRequired,
     }).isRequired,
-    entityData: React.PropTypes.object.isRequired,
-    onEdit: React.PropTypes.func.isRequired,
-    onRemove: React.PropTypes.func.isRequired,
+    entityData: PropTypes.object.isRequired,
+    onEdit: PropTypes.func.isRequired,
+    onRemove: PropTypes.func.isRequired,
 };
 
 export default Tooltip;
